Feat: refactored to make modular
