--- album.json
@@ -1,26 +1,0 @@
-[
-	{
-		"id": 304,
-		"publish_date": "2022-02-02",
-		"on_spotify": false,
-		"genre": "Afrobeats"
-	},
-	{
-		"id": 317,
-		"publish_date": "2021-02-03",
-		"on_spotify": false,
-		"genre": "Hip-Hop"
-	},
-	{
-		"id": 852,
-		"publish_date": "2019-04-05",
-		"on_spotify": true,
-		"genre": "_frobeats"
-	},
-	{
-		"id": 50,
-		"publish_date": "2022-03-03",
-		"on_spotify": true,
-		"genre": "Comedy"
-	}
-]
