+++ classes/music_album.rb
