--- games.json
@@ -1,1 +1,0 @@
-[{"id":543,"multiplayer":"ahmed","last_played":"3","publish_date":"2015-02-20","author":{"first_name":"test","last_name":"fest"}},{"id":801,"multiplayer":"kevin","last_played":"1","publish_date":"2020-05-10","author":{"first_name":"fizz","last_name":"buzz"}},{"id":598,"multiplayer":"Anselem","last_played":"3","publish_date":"2015-02-05","author":{"first_name":"tancent","last_name":"development"}}]
